feat: comprehensive billing system enhancements
Migration Improvements:
- Consolidate billing migrations into 4 core files for better organization
- Move legacy migrations to Backup folder for reference
- Add usage changes, quotes, and quote line items migrations
- Add non-profit fields, pricing tiers, and billing frequency support
- Add billing invitations system for customer onboarding

Seeding Enhancements:
- Add InvoiceSeeder with 3-8 invoices per company, multiple statuses
- Add PaymentSeeder with multiple payment methods (stripe, ach, wire, check, cash)
- Add PriceOverrideSeeder with fixed/discount/markup pricing types
- Add CreditNoteSeeder for invoice credits and refunds
- Add DisputeSeeder for invoice dispute tracking
- Add UsageChangeSeeder for subscription quantity approvals
- Add VolumePricingExampleSeeder for tiered pricing demos
- Update existing seeders with enhanced data generation

Model & Controller Updates:
- Enhance Invoice, Payment, Quote models with better relationships
- Add Dispute, Contract, Invitation models
- Update controllers with improved authorization and validation
- Add public quote acceptance/rejection flows
- Enhance executive dashboard with better metrics

UI/UX Improvements:
- Add tabbed interfaces for finance views (invoices, quotes, payments)
- Implement invoice timeline component
- Add enhanced pre-flight checklist view
- Create reports hub and settings hub views
- Improve quote creation/editing with pricing tier support
- Add technician feedback interface
- Enhance portal dashboard with better navigation

Service Layer:
- Update PricingEngineService with tax and credit handling
- Enhance InvoiceGenerator with better line item management
- Improve QuoteConversionService for quote-to-subscription flow
- Update integration services (Xero, Stripe, PayPal, GoCardless)
- Add AlertService, AnalyticsService, ClientHealthService

Documentation:
- Add quote pricing quick reference and tier documentation
- Document tabbed views implementation
- Create UI transformation summaries
- Add world-class UX enhancements guide

+++ Services/AlertService.php
@@ -12,6 +12,8 @@ class AlertService
 {
     /**
      * Check all thresholds and return triggered alerts.
+     *
+     * @return array<int, array<string, mixed>>
      */
     public function checkThresholds(): array
     {
@@ -88,6 +90,8 @@ class AlertService
 
     /**
      * Get threshold configuration.
+     *
+     * @return array<string, array{enabled: bool, threshold: float}>
      */
     public function getThresholdConfig(): array
     {
@@ -135,6 +139,9 @@ class AlertService
 
     /**
      * Send an alert through configured channels.
+     *
+     * @param string $alertType
+     * @param array<string, mixed> $data
      */
     public function sendAlert(string $alertType, array $data): void
     {
@@ -175,7 +182,7 @@ class AlertService
      */
     protected function calculateArTotal(): float
     {
-        return Invoice::whereIn('status', ['sent', 'overdue'])
+        return (float) Invoice::whereIn('status', ['sent', 'overdue'])
             ->sum('total');
     }
 
@@ -184,12 +191,12 @@ class AlertService
      */
     protected function calculateMrrChange(): float
     {
-        $currentMrr = Subscription::where('is_active', true)
+        $currentMrr = (float) Subscription::where('is_active', true)
             ->where('billing_frequency', 'monthly')
             ->sum('effective_price');
 
         // Get last month's MRR from cache or calculate
-        $lastMonthMrr = Cache::get('billing_mrr_last_month', $currentMrr);
+        $lastMonthMrr = (float) Cache::get('billing_mrr_last_month', $currentMrr);
 
         if ($lastMonthMrr == 0) {
             return 0.0;
@@ -205,6 +212,9 @@ class AlertService
 
     /**
      * Get companies with margins below threshold.
+     *
+     * @param float $threshold
+     * @return \Illuminate\Database\Eloquent\Collection<int, Company>
      */
     protected function getLowMarginClients(float $threshold)
     {
